Deep-Q Network Algorithm
The algorithm monitors the workload of an apache server, performs parameter changes, compares it with the SLA of the website.
If the performance is better, a better reward is yielded and the Q-Tables are updated accordingly during the training phase
While testing, the built Q-table is used to fetch the best configuration based on the Q-values(max), and the performance benchmarking is performed against those values. If the performance is better, the new reward values are updated in the Q-table for future use.